requirements.txt
matplotlib==3.8.3
opencv-python==4.9.0.80
numpy==1.21.5
numpy-stl==3.1.1
ipykernel==6.29.3
pypylon==3.0.1
pyserial
